+++ examples/omeinsum_julia/omeinsum_treesa_optimizer.py
@@ -4,7 +4,7 @@ from typing import List, Set, Dict, Tuple
 class OMEinsumTreeSAOptimizer(object):
     def __init__(
         self,
-        sc_target: float = 20,
+        sc_target: int = 20,
         betas: Tuple[float, float, float] = (0.01, 0.01, 15),
         ntrials: int = 10,
         niters: int = 50,
